[waicolle] give up creating thumbnail for ridiculous trades
it takes too long to generate and probably wouldn’t need it.

the thumbnail url was nice because it could be created asynchronously
after the embed is sent. trying to upload it to producer takes that
away and instead we time out waiting for the collage to be generated.

we could try compressing the ids in the url to work around the 2048
characters limitation but it is rare enough that we don’t have to waste
time on implementing that (or another crazy workaround).