don’t set orjson numpy serialize option
The last remaining use of numpy is for the RNG (where it should still
make sense compared to the default random module in python)
We don’t use numpy to generate data that is sent to our server so we can
drop this option.

honestly wanted to see if we could drop orjson but it’s useful to
serialize dataclasses and similar objects.