[waicolle] fix casting autocompletion when user is too fast™
if user sends a command before the completion kicks in some commands can
fail because completion replaces the argument of the user into another
value.

so this commit create a helper that can call the completion function to
get the value when needed.

(probably could be used in more places)