remove dependency on toolz
toolz won’t provide typing any time soon.
most functions are calling back to stdlib/itertools functions.
some other can be replicated with lambda function, which one looks
better is in the eye of the beholder.

the only function without an obvious counterpart is interleave, so it is
now vendored with the appropriate license baked in.