homebrew saucenao client
so at least we don’t have to deal with a library breaking its API for
some reason. the SauceNao API seems not obvious to use generically for
all its index which would be interesting to put in a library, but we
already implement custom logic in nanachan depending on the index so its
whatever.

would be good to use the danbooru index or others for lookups as the
pixiv index seems to be late.